quote_generation/requirements.txt
Flask==3.1.0
requests==2.32.3
numpy==2.2.1
azure-ai-inference==1.0.0b7